Ajustes apontados pela análise estática

+++ tests/BladeEngineTest.php
@@ -66,7 +66,7 @@ class BladeEngineTest extends TestCase
         $listFiles = $this->listFiles($cachePath);
 
         foreach ($listFiles as $file) {
-            $permission = fileperms($file);
+            $permission = (int) fileperms($file);
             $permissionOctal = decoct($permission);
             $permissionQuad = substr($permissionOctal, -4);
